Refactoring for Sync and WaitSync functions.
- Add a new file sync.h/cc in mojo public.
- Move Sync and WaitSync functions from platform_handle.h/cc
  to sync.h/cc.
- Rename SyncSharedMemoryHandle to SyncSharedMemory.
