+++ cc/raster/one_copy_raster_buffer_provider.cc
@@ -31,7 +31,7 @@
 
 #if defined(CASTANETS)
 #include "base/command_line.h"
-#include "mojo/public/cpp/system/platform_handle.h"
+#include "mojo/public/cpp/system/sync.h"
 #endif
 
 namespace cc {
@@ -362,7 +362,7 @@ void OneCopyRasterBufferProvider::PlaybackToStagingBuffer(
 #if defined(CASTANETS)
     if (std::string("renderer") ==
         base::CommandLine::ForCurrentProcess()->GetSwitchValueASCII("type")) {
-      mojo::SyncSharedMemoryHandle(
+      mojo::SyncSharedMemory(
           buffer->GetHandle().handle.GetGUID(), 0,
           staging_buffer->size.width() * staging_buffer->size.height() * 4);
     }
