+++ gpu/command_buffer/service/common_decoder.cc
@@ -13,7 +13,7 @@
 #include "gpu/command_buffer/service/command_buffer_service.h"
 
 #if defined(CASTANETS)
-#include "mojo/public/cpp/system/platform_handle.h"
+#include "mojo/public/cpp/system/sync.h"
 #endif
 
 namespace gpu {
@@ -342,8 +342,8 @@ error::Error CommonDecoder::HandleGetBucketStart(
 #if defined(CASTANETS)
   scoped_refptr<gpu::Buffer> buffer =
       command_buffer_service_->GetTransferBuffer(data_memory_id);
-  mojo::SyncSharedMemoryHandle(buffer->backing()->GetGUID(), data_memory_offset,
-                               data_memory_size);
+  mojo::SyncSharedMemory(buffer->backing()->GetGUID(), data_memory_offset,
+                         data_memory_size);
 #endif
   return error::kNoError;
 }
@@ -372,7 +372,7 @@ error::Error CommonDecoder::HandleGetBucketData(uint32_t immediate_data_size,
 #if defined(CASTANETS)
   scoped_refptr<gpu::Buffer> buffer =
       command_buffer_service_->GetTransferBuffer(args.shared_memory_id);
-  mojo::SyncSharedMemoryHandle(buffer->backing()->GetGUID(), offset, size);
+  mojo::SyncSharedMemory(buffer->backing()->GetGUID(), offset, size);
 #endif
   return error::kNoError;
 }
@@ -385,8 +385,7 @@ error::Error CommonDecoder::HandleSyncResultData(
       *static_cast<const volatile cmd::SyncResultData*>(cmd_data);
   scoped_refptr<gpu::Buffer> buffer =
       command_buffer_service_->GetTransferBuffer(args.id);
-  mojo::SyncSharedMemoryHandle(buffer->backing()->GetGUID(), args.offset,
-                               args.size);
+  mojo::SyncSharedMemory(buffer->backing()->GetGUID(), args.offset, args.size);
 
   // Set result data to '0' for use in the next command on Castanets.
   // Some apis should set data to non-zero, it can cause an issue.
