+++ media/renderers/video_resource_updater.cc
@@ -48,7 +48,7 @@
 #include "ui/gl/trace_util.h"
 
 #if defined(CASTANETS)
-#include "mojo/public/cpp/system/platform_handle.h"
+#include "mojo/public/cpp/system/sync.h"
 #endif
 
 namespace media {
@@ -873,8 +873,10 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
         video_renderer_->Copy(video_frame, &canvas, Context3D());
 #if defined(CASTANETS)
         // Compute the memory size with a RGBA_8888 format.
-        size_t memory_bytes = software_resource->resource_size().width() * software_resource->resource_size().height() * 4;
-        mojo::SyncSharedMemoryHandle(software_resource->GetSharedMemoryGuid(), 0, memory_bytes);
+        size_t memory_bytes = software_resource->resource_size().width() *
+                              software_resource->resource_size().height() * 4;
+        mojo::SyncSharedMemory(software_resource->GetSharedMemoryGuid(), 0,
+                               memory_bytes);
 #endif
       } else {
         HardwarePlaneResource* hardware_resource = plane_resource->AsHardware();
