+++ mojo/public/cpp/system/BUILD.gn
@@ -2,6 +2,8 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
+import("//build/config/features.gni")
+
 # Deletes libsystem.dylib from the build dir, since it shadows
 # /usr/lib/libSystem.dylib on macOS.
 # TODO(thakis): Remove this after a while.
@@ -61,6 +63,13 @@ component("system") {
     "wait_set.cc",
     "wait_set.h",
   ]
+
+  if (enable_castanets) {
+    sources += [
+      "sync.cc",
+      "sync.h",
+    ]
+  }
 
   public_deps = [
     "//base",
