+++ mojo/public/cpp/system/platform_handle.cc
@@ -13,10 +13,6 @@
 #include "base/mac/mach_logging.h"
 #endif
 
-#if defined(CASTANETS)
-#include "base/memory/shared_memory_tracker.h"
-#endif
-
 namespace mojo {
 
 namespace {
@@ -359,40 +355,6 @@ MojoResult UnwrapSharedMemoryHandle(
 
   return MOJO_RESULT_OK;
 }
-
-#if defined(CASTANETS)
-MojoResult SyncSharedMemoryHandle(const base::UnguessableToken& guid,
-                                  size_t offset,
-                                  size_t sync_size) {
-  MojoSharedBufferGuid mojo_guid;
-  mojo_guid.high = guid.GetHighForSerialization();
-  mojo_guid.low = guid.GetLowForSerialization();
-
-  return MojoSyncPlatformSharedMemoryRegion(
-      &mojo_guid, offset, sync_size);
-}
-
-MojoResult SyncSharedMemoryHandle2d(const base::UnguessableToken& guid,
-                                    size_t offset,
-                                    size_t sync_size,
-                                    size_t width,
-                                    size_t stride) {
-  MojoSharedBufferGuid mojo_guid;
-  mojo_guid.high = guid.GetHighForSerialization();
-  mojo_guid.low = guid.GetLowForSerialization();
-
-  return MojoSyncPlatformSharedMemoryRegion2d(&mojo_guid, offset, sync_size,
-                                              width, stride);
-}
-
-MojoResult WaitSyncSharedMemory(const base::UnguessableToken& guid) {
-  MojoSharedBufferGuid mojo_guid;
-  mojo_guid.high = guid.GetHighForSerialization();
-  mojo_guid.low = guid.GetLowForSerialization();
-
-  return MojoWaitSyncPlatformSharedMemoryRegion(&mojo_guid);
-}
-#endif
 
 ScopedSharedBufferHandle WrapReadOnlySharedMemoryRegion(
     base::ReadOnlySharedMemoryRegion region) {
